Moved entirely over to BlockReader w/out templates
